requirements.txt
flask
flask_sqlalchemy
gunicorn
requests
python-telegram-bot